+++ cameo.py
@@ -1,15 +1,19 @@
 #!/usr/bin/env python3
-
-import binascii
-from construct import *
-import cv2
-import fcntl
+import argparse
 import logging
+
+import cv2
 import numpy
-import time
+from construct import *
+import fcntl
 import sys
 
-from filters import *
+from filters.filter_add_text import FilterAddText
+from filters.filter_background_blur import FilterBackgroundBlur
+from filters.filter_video import FilterVideo
+from filters.filter_blur import FilterBlur
+from filters.filter_add_image import FilterAddImage
+from filters.filter_color import FilterColor
 
 V4L2_FIELD_NONE = 1
 V4L2_BUF_TYPE_VIDEO_OUTPUT = 2
@@ -78,12 +82,12 @@ def open_video_out(camera_out, width, height):
 def main(camera_in=0, camera_out=1, do_flip=False, thumbnail=False):
     current_filter = None
     keys = {
-        " ": (FilterColor,    [ (255, 0, 211) ]),
-        ")": (FilterAddImage, [ "img/smile.png" ]),
-        "a": (FilterAddImage, [ "img/applause.png" ]),
-        "t": (FilterAddText,  [ "(be right back)" ]),
-        "v": (FilterVideo,    [ "img/rick-astley-never-gonna-give-you-up-video.mp4" ]),
-        "b": (FilterBlur ,    []),
+        " ": (FilterColor,          [ (255, 0, 211) ]),
+        ")": (FilterAddImage,       [ "img/smile.png" ]),
+        "a": (FilterAddImage,       [ "img/applause.png" ]),
+        "t": (FilterAddText,        [ "(be right back)" ]),
+        "b": (FilterBlur,           []),
+        "v": (FilterBackgroundBlur, []),
     }
 
     capture = open_capture(camera_in)
